refactor: dedicated test classes for predefined text preprocessors
