+++ testing/camunda-process-test-spring/src/test/java/io/camunda/process/test/api/SemanticSimilarityConfigBootstrapIT.java
@@ -327,7 +327,7 @@ public class SemanticSimilarityConfigBootstrapIT {
 
     @Test
     void shouldThrowWhenRequiredFieldMissing() {
-      final ProviderConfig config = new BaseProviderConfig.OpenAiConfig(null, null);
+      final ProviderConfig config = new BaseProviderConfig.OpenAiConfig(null, null, null, null);
       final OpenAiEmbeddingModelAdapterProvider provider =
           new OpenAiEmbeddingModelAdapterProvider();
       assertThatThrownBy(() -> provider.create(config)).isInstanceOf(IllegalStateException.class);
